KS-1-4 error fixed

+++ src/main/kotlin/lesson_1/task_4.kt
@@ -4,7 +4,7 @@ fun main() {
     val length: Long = 40868600000L
     val age: Byte = 27
     val partOfDay: Float = 0.075f
-    val seconds: Int = 6480
+    val seconds: Short = 6480
     val partOfYear: Double = 2.0547945205479453E-4
     val apogee: Int = 327000
